Validate utf codepoint at full width, not after 32-bit truncation
<<N/utf8>> (and utf16/utf32) must raise badarg when N is not a valid
Unicode code point. AtomVM rejected most invalid values but silently
encoded a family of large negative integers: the opcode handlers and JIT
primitives computed a 64-bit avm_int_t src_value, then passed it to
encoders whose codepoint parameter was uint32_t (and to the JIT
primitives, whose parameter was int). The value was implicitly truncated
to its low 32 bits before is_invalid_codepoint(int32_t) ever saw it. A
value such as -2^59 (0xF800_0000_0000_0000) truncates to 0 -- a valid
codepoint -- so <<(-576460752303423488)/utf8>> produced <<0>> where BEAM
raises badarg; -2^59+65 produced <<"A">>. This affected both the emulator
and JIT builds (erlfuzz bug #16, utf_negative_bignum_bypass).

Widen the codepoint all the way through: is_invalid_codepoint now takes
int64_t, the bitstring_utf{8,16,32}_encode / _size / insert_utf{8,16,32}
functions take int64_t c, and the jit.c primitives + jit.h interface
pointers match. The full-width value now reaches the range check, so
out-of-range negatives (and any value outside [0, 0x10FFFF]) are rejected
on both builds. The encoders only dereference c after the range gate, so
the narrower arithmetic inside them is unchanged. No Erlang codegen
change is needed -- the JIT already passes a full register-width value;
only the C declared parameter types were narrowing it.

Regression test in test_bs_utf: put_utf8/16/32 now assert badarg for
-2^59 and -2^59+65 (low 32 bits 0 and 65). Verified on JIT and emulator.

+++ src/libAtomVM/bitstring.h
@@ -293,7 +293,7 @@ static inline bool bitstring_insert_integer(term dst_bin, size_t offset, avm_int
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-bool bitstring_utf8_encode(uint32_t c, uint8_t *buf, size_t *out_size);
+bool bitstring_utf8_encode(int64_t c, uint8_t *buf, size_t *out_size);
 
 /**
  * @brief Encode a character to UTF-16.
@@ -306,7 +306,7 @@ bool bitstring_utf8_encode(uint32_t c, uint8_t *buf, size_t *out_size);
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-bool bitstring_utf16_encode(uint32_t c, uint8_t *buf, enum BitstringFlags bs_flags, size_t *out_size);
+bool bitstring_utf16_encode(int64_t c, uint8_t *buf, enum BitstringFlags bs_flags, size_t *out_size);
 
 /**
  * @brief Decode a character from UTF-16.
@@ -331,7 +331,7 @@ bool bitstring_utf16_decode(const uint8_t *buf, size_t len, int32_t *c, size_t *
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-bool bitstring_utf32_encode(uint32_t c, uint8_t *buf, enum BitstringFlags bs_flags);
+bool bitstring_utf32_encode(int64_t c, uint8_t *buf, enum BitstringFlags bs_flags);
 
 /**
  * @brief Decode a character from UTF-32.
@@ -354,7 +354,7 @@ bool bitstring_utf32_decode(const uint8_t *buf, size_t len, int32_t *c, enum Bit
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-static inline bool bitstring_utf8_size(uint32_t c, size_t *out_size)
+static inline bool bitstring_utf8_size(int64_t c, size_t *out_size)
 {
     return bitstring_utf8_encode(c, NULL, out_size);
 }
@@ -367,7 +367,7 @@ static inline bool bitstring_utf8_size(uint32_t c, size_t *out_size)
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-static inline bool bitstring_utf16_size(uint32_t c, size_t *out_size)
+static inline bool bitstring_utf16_size(int64_t c, size_t *out_size)
 {
     return bitstring_utf16_encode(c, NULL, BigEndianInteger, out_size);
 }
@@ -382,7 +382,7 @@ static inline bool bitstring_utf16_size(uint32_t c, size_t *out_size)
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-static inline bool bitstring_insert_utf8(term dst_bin, size_t offset, uint32_t c, size_t *out_size)
+static inline bool bitstring_insert_utf8(term dst_bin, size_t offset, int64_t c, size_t *out_size)
 {
     // size was verified by a bs_utf8_size instruction call
     uint8_t *dst = (uint8_t *) term_binary_data(dst_bin) + (offset >> 3);
@@ -417,7 +417,7 @@ static inline bool bitstring_match_utf8(term src_bin, size_t offset, uint32_t *c
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-static inline bool bitstring_insert_utf16(term dst_bin, size_t offset, uint32_t c, enum BitstringFlags bs_flags, size_t *out_size)
+static inline bool bitstring_insert_utf16(term dst_bin, size_t offset, int64_t c, enum BitstringFlags bs_flags, size_t *out_size)
 {
     // size was verified by a bs_utf8_size instruction call
     uint8_t *dst = (uint8_t *) term_binary_data(dst_bin) + (offset >> 3);
@@ -452,7 +452,7 @@ static inline bool bitstring_match_utf16(term src_bin, size_t offset, int32_t *c
  * @return \c true if encoding was successful, \c false if c is not a valid
  * unicode character
  */
-static inline bool bitstring_insert_utf32(term dst_bin, size_t offset, uint32_t c, enum BitstringFlags bs_flags)
+static inline bool bitstring_insert_utf32(term dst_bin, size_t offset, int64_t c, enum BitstringFlags bs_flags)
 {
     uint8_t *dst = (uint8_t *) term_binary_data(dst_bin) + (offset >> 3);
     return bitstring_utf32_encode(c, dst, bs_flags);
